UDPClient.SendAsync is used now to send a data instead of Task.Run(()=>UDPClient.Send())

+++ LANPaint/Services/UDP/INetworkBroadcaster.cs
@@ -6,7 +6,7 @@ namespace LANPaint.Services.UDP
     public interface INetworkBroadcaster : IDisposable
     {
         Task<byte[]> ReceiveAsync();
-        Task<long> SendAsync(byte[] bytes);
+        Task<int> SendAsync(byte[] bytes);
         ValueTask ClearBufferAsync();
     }
 }
